backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
asyncpg
alembic
psycopg2-binary
requests
openai-agents
python-dateutil
APScheduler
pywebpush